fixes for https://github.com/openzfsonosx/zfs/issues/598
These fixes make a DEBUG zfs.kext and zdb work.

abd.c API differs among ZOL, Illumos and O3X (the
latter two are the closest).

In particular abd_copy() no longer takes a size argument.

For O3X and Illumos it seems like bad practice to use
abd_copy() when the source and destination abds are of
dissimilar size.  O3X will ASSERT in this case, and in the
others range-checking is poor.  There are similar problems
for abd_copy_from_buf() and others.

Instead, when two ABDs are dissimilar, one should do a
range-checking with ASSERTs above the call site and then
use the _off() variants.  For example:

ASSERT3S(srcabd->abd_size, >=, copysize);
ASSERT3S(dstabd->abd_size, >=, copysize);
abd_copy_off(srcabd, dstabd, 0, 0, copysize);

This makes the intention that there will be uncopied bytes
left in one or the other ABD.

abd.h and abd.c functions can use some DEBUG
range-checking ASSERTs in the *_off() functions.

Add in further defensive ASSERTs in abd.c.

This also squashes in
d0be996bb6e41f9cd3d45ceae59bc511a7778126 Random corrections

0d467f47acb4e987cdefef286c2cccf85284de20 Minor vdev_file.c cleanup

678ff0b915f2220ba0269fc36bae32cd87e83d88 Correct abd usage in vdev_raidz.c

bb4a19210eb7ce1f46f0adc045b61b8b34a27acf Finalised vdev_raidz.c

+++ ZFSin/zfs/include/sys/abd.h
@@ -119,8 +119,8 @@ abd_copy(abd_t *dabd, abd_t *sabd, size_t size)
 	ASSERT3P(sabd,!=,NULL);
 	ASSERT3P(dabd,!=,sabd);
 	if (dabd != sabd) {
-		ASSERT3S(dabd->abd_size,==,size);
-		ASSERT3S(sabd->abd_size,==,size);
+		ASSERT3S((size_t)dabd->abd_size,==,size);
+		ASSERT3S((size_t)sabd->abd_size,==,size);
 		abd_copy_off(dabd, sabd, 0, 0, size);
 	}
 }
@@ -128,28 +128,28 @@ abd_copy(abd_t *dabd, abd_t *sabd, size_t size)
 static inline void
 abd_copy_from_buf(abd_t *abd, void *buf, size_t size)
 {
-	ASSERT3S(abd->abd_size,==,size);
+	ASSERT3S((size_t)abd->abd_size,==,size);
 	abd_copy_from_buf_off(abd, buf, 0, size);
 }
 
 static inline void
 abd_copy_to_buf(void* buf, abd_t *abd, size_t size)
 {
-	ASSERT3S(abd->abd_size,==,size);
+	ASSERT3S((size_t)abd->abd_size,==,size);
 	abd_copy_to_buf_off(buf, abd, 0, size);
 }
 
 static inline int
 abd_cmp_buf(abd_t *abd, void *buf, size_t size)
 {
-	ASSERT3S(abd->abd_size,==,size);
+	ASSERT3S((size_t)abd->abd_size,==,size);
 	return (abd_cmp_buf_off(abd, buf, 0, size));
 }
 
 static inline void
 abd_zero(abd_t *abd, size_t size)
 {
-	ASSERT3S(abd->abd_size,==,size);
+	ASSERT3S((size_t)abd->abd_size,==,size);
 	abd_zero_off(abd, 0, size);
 }
